"first pass at factoring out Clocks code; updates #1171"
